Moved NotFound to its own component
